reduce inheritance depth
